Packaging: fix service start issue due to package not found
setup.py fetches content of requirements.txt and omits the
last character if the file does not ends with new line.
This causes DistributionNotFound error if the requirements.txt
file does not ends with a new line.

This patch fixes this issue by just strip out the unnecessary
space and newline only if it exists.

tendrl-bug-id: Tendrl/performance-monitoring#58
Signed-off-by: Timothy Asir J <tjeyasin@redhat.com>