Fix unwanted re-generation of collectd plugins
Fix unwanted re-generation of collectd plugins across performance-monitoring
restarts. Prior to this approach the nodes on which configurations are made
were stored in-memory to avoid regeneration of configurations during the next
scan through the list of clusters and nodes. But the problem with that approach
is that if performance-monitoring is restarted for any reason such as node goes
down and hence performance-monitoring was started back once the node is back up,
or for any other reason, the configurations would be regenerated as the in-memory
data about the configurations made were lost.

Hence this PR puts that in-memory state info about the configurations made to
etcd which is persisted across restarts and hence configuration regenerations
are avoided.

tendrl-bug-id: Tendrl/performance-monitoring#202
Signed-off-by: anmolbabu <anmolbudugutta@gmail.com>